+++ src/superio/winbond/w83627hf/w83627hf.h
@@ -3,7 +3,7 @@
  *
  * Copyright (C) 2000 AG Electronics Ltd.
  * Copyright (C) 2003-2004 Linux Networx
- * Copyright (C) 2004 Tyan By LYH change from PC87360
+ * Copyright (C) 2004 Tyan
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,16 +24,16 @@
 #define SUPERIO_WINBOND_W83627HF_W83627HF_H
 
 #define W83627HF_FDC              0   /* Floppy */
-#define W83627HF_PP               1   /* Parallel Port */
+#define W83627HF_PP               1   /* Parallel port */
 #define W83627HF_SP1              2   /* Com1 */
 #define W83627HF_SP2              3   /* Com2 */
-#define W83627HF_KBC              5   /* Keyboard & Mouse */
+#define W83627HF_KBC              5   /* PS/2 keyboard & mouse */
 #define W83627HF_CIR              6
 #define W83627HF_GAME_MIDI_GPIO1  7
 #define W83627HF_GPIO2            8
 #define W83627HF_GPIO3            9
 #define W83627HF_ACPI            10
-#define W83627HF_HWM             11   /* Hardware Monitor */
+#define W83627HF_HWM             11   /* Hardware monitor */
 
 /* #define W83627HF_GPIO_DEV PNP_DEV(0x2e, W83627HF_GPIO) */
 /* #define W83627HF_XBUS_DEV PNP_DEV(0x2e, W83627HF_XBUS) */
@@ -115,8 +115,8 @@
 
 #if defined(__PRE_RAM__) && !defined(__ROMCC__)
 void w83627hf_disable_dev(device_t dev);
-void w83627hf_enable_dev(device_t dev, unsigned iobase);
-void w83627hf_enable_serial(device_t dev, unsigned iobase);
+void w83627hf_enable_dev(device_t dev, u16 iobase);
+void w83627hf_enable_serial(device_t dev, u16 iobase);
 void w83627hf_set_clksel_48(device_t dev);
 #endif
